Refactor adapter
